requirements.txt
torch
torchvision
torchaudio
transformers
diffusers
accelerate
opencv-python
scikit-image
pandas
numpy
tqdm
Pillow
requests
huggingface_hub
peft
scikit-learn
tabulate
matplotlib
